double roundOff function added and gas type updated as long

+++ app/views/companion/blockchain/AddZone.scala
@@ -15,6 +15,6 @@ object AddZone {
     )(Data.apply)(Data.unapply)
   )
 
-  case class Data(from: String, to: String, zoneID: String, gas: Int, mode: String, password: String)
+  case class Data(from: String, to: String, zoneID: String, gas: Long, mode: String, password: String)
 
 }
