+++ app/views/companion/blockchain/SetBuyerFeedback.scala
@@ -16,6 +16,6 @@ object SetBuyerFeedback {
     )(Data.apply)(Data.unapply)
   )
 
-  case class Data(from: String, to: String, pegHash: String, rating: Int, gas: Int, mode: String, password: String)
+  case class Data(from: String, to: String, pegHash: String, rating: Int, gas: Long, mode: String, password: String)
 
 }
